// src/shared/enums/amenity.ts
export enum EAmenities {
  Wifi = "Wi-Fi",
  AirConditioning = "Ar-condicionado",
  Kitchen = "Cozinha",
  Parking = "Estacionamento",
  Pool = "Piscina",
  Grill = "Churrasqueira",
  Fireplace = "Lareira",
  Bath = "Hidromassagem",
  Heater = "Aquecedor",
  EquippedKitchen = "Cozinha equipada",
  CompactKitchen = "Cozinha compacta"
}